+++ src/lib/format/config/tsconfig.ts
@@ -12,7 +12,7 @@ const CACHE = new NodeCache({ stdTTL: 5 });
 
 export function loadTsConfig(fileName: string, configPath?: string) {
   const log = logger('format-imports.loadTsConfig');
-  log.debug('Loading TS config for fileName:', fileName, 'from', configPath ?? 'default');
+  log.debug('Loading TS config for:', fileName, 'from', configPath ?? 'default');
   try {
     const { findConfigFile, readConfigFile, parseJsonConfigFileContent, sys, version } =
       requireModule('typescript', fileName, ts);
@@ -21,14 +21,14 @@ export function loadTsConfig(fileName: string, configPath?: string) {
     if (!configFile) return undefined;
     const opt = CACHE.get(configFile);
     if (isObject(opt)) return opt as CompilerOptions;
-    log.debug('Loading TS config from:', configFile);
     const { config } = readConfigFile(configFile, sys.readFile.bind(sys));
     const { options } = parseJsonConfigFileContent(config, sys, path.dirname(configFile));
     CACHE.set(configFile, options);
+    log.debug('Loading TS config from', configFile, 'and options:', options);
     return options;
   } catch (e: unknown) {
     const msg = e instanceof Error ? e.message : `${e}`;
-    log.warn('Failed to load TS config for fileName:', fileName, 'with error:', msg);
+    log.warn('Failed to load TS config for', fileName, 'with error:', msg);
     return undefined;
   }
 }