+++ src/test/lib/examples/assert/origin.ts
@@ -1,6 +1,0 @@
-import A from 'a' assert {  }
-import D from 'a' assert { a: '' }
-import B from 'b' assert { ab:    'xyz' }
-import C from 'c' assert { a: 'x',bc:'xyz' }
-
-export {A, B, C,D}