+++ src/test/lib/examples/attributes/assert.result.ts
@@ -1,0 +1,30 @@
+import A from 'a';
+import D from 'a' assert { a: '' };
+import B, {
+  debounce,
+  isEmpty,
+  omitBy,
+} from 'b' assert { ab: 'xyz' };
+import E from 'b' with { ab: 'xyz' };
+import C from 'c' assert {
+  a: 'x',
+  bc: 'xyz',
+  asdfads: 'adg',
+  gaefasd: 'gege',
+  gaeaesadsasdf: 'agefasdfasdgasdf',
+};
+import {
+  OnboardingStatusFilter,
+  PaymentsStatusFilter,
+} from 'd' assert {
+  a: 'x',
+  bc: 'xyz',
+  asdfads: 'adg',
+  gaefasd: 'gege',
+  gaeaesadsasdf: 'agefasdfasdgasdf',
+};
+
+export {A, B, C,D,E,debounce,
+  isEmpty,
+  omitBy, OnboardingStatusFilter,
+  PaymentsStatusFilter}