+++ src/test/lib/examples/attributes/with.origin.ts
@@ -1,0 +1,26 @@
+import A from 'a' with {  }
+import D from 'a' with { a: '' }
+import B from 'b' with { ab:    'xyz' }
+import C from 'c' with { a: 'x',bc:'xyz', asdfads: "adg",
+gaefasd: "gege",
+gaeaesadsasdf: "agefasdfasdgasdf", }
+import E from 'b' assert { ab:    'xyz' }
+
+import {debounce,
+  isEmpty,
+  omitBy,} from 'b' with { ab:    'xyz' }
+
+
+export {A, B, C,D,E,debounce,
+  isEmpty,
+  omitBy, OnboardingStatusFilter,
+  PaymentsStatusFilter}
+
+  import {  OnboardingStatusFilter,
+    PaymentsStatusFilter,} from 'c' with {
+    a: 'x',
+    bc: 'xyz',
+    asdfads: 'adg',
+    gaefasd: 'gege',
+    gaeaesadsasdf: 'agefasdfasdgasdf1',
+  };